chore: StoryName 타입을 domain으로 이동
