+++ packages/domain/types/chat.ts
@@ -4,3 +4,10 @@ export interface Message {
   msg: string;
   profileImg: string;
 }
+
+export enum StoryName {
+  EXECUTION = 'EXECUTION',
+  PUBLISH_VICTIM = 'PUBLISH_VICTIM',
+  NO_KILL = 'NO_KILL',
+  PUBLISH_SURVIVOR = 'PUBLISH_SURVIVOR',
+}
